@@ -1,73 +1,84 @@
 module github.com/velour/catbase
 
-go 1.16
+go 1.18
 
 require (
 	code.chrissexton.org/cws/getaoc v0.0.0-20191201043947-d5417d4b618d
-	github.com/AlekSi/pointer v1.1.0 // indirect
 	github.com/ChimeraCoder/anaconda v2.0.0+incompatible
-	github.com/ChimeraCoder/tokenbucket v0.0.0-20131201223612-c5a927568de7 // indirect
 	github.com/PuerkitoBio/goquery v1.5.0
+	github.com/bwmarrin/discordgo v0.23.2
+	github.com/cdipaolo/goml v0.0.0-20190412180403-e1f51f713598
+	github.com/chrissexton/leftpad v0.0.0-20181207133115-1e93189d2fff
+	github.com/chrissexton/sentiment v0.0.0-20190927141846-d69c422ba035
+	github.com/fogleman/gg v1.2.1-0.20190220221249-0403632d5b90
+	github.com/go-chi/chi/v5 v5.0.3
+	github.com/gocolly/colly v1.2.0
+	github.com/google/uuid v1.1.1
+	github.com/itchyny/gojq v0.12.3
+	github.com/james-bowman/nlp v0.0.0-20191016091239-d9dbfaff30c6
+	github.com/jmoiron/sqlx v1.2.0
+	github.com/kevinburke/twilio-go v0.0.0-20200424172635-4f0b2357b852
+	github.com/mattn/go-sqlite3 v1.14.8
+	github.com/mmcdole/gofeed v1.0.0-beta2
+	github.com/nfnt/resize v0.0.0-20180221191011-83c6a9932646
+	github.com/olebedev/when v0.0.0-20190311101825-c3b538a97254
+	github.com/rs/zerolog v1.15.0
+	github.com/slack-go/slack v0.9.4
+	github.com/stretchr/testify v1.4.0
+	github.com/trubitsyn/go-zero-width v1.0.1
+	github.com/velour/velour v0.0.0-20160303155839-8e090e68d158
+	golang.org/x/crypto v0.0.0-20211117183948-ae814b36b871
+	gopkg.in/go-playground/webhooks.v5 v5.13.0
+)
+
+require (
+	github.com/AlekSi/pointer v1.1.0 // indirect
+	github.com/ChimeraCoder/tokenbucket v0.0.0-20131201223612-c5a927568de7 // indirect
 	github.com/andybalholm/cascadia v1.1.0 // indirect
 	github.com/antchfx/htmlquery v1.2.0 // indirect
 	github.com/antchfx/xmlquery v1.2.0 // indirect
 	github.com/antchfx/xpath v1.1.1 // indirect
 	github.com/armon/go-radix v1.0.0 // indirect
 	github.com/azr/backoff v0.0.0-20160115115103-53511d3c7330 // indirect
-	github.com/bwmarrin/discordgo v0.23.2
-	github.com/cdipaolo/goml v0.0.0-20190412180403-e1f51f713598
-	github.com/chrissexton/leftpad v0.0.0-20181207133115-1e93189d2fff
-	github.com/chrissexton/sentiment v0.0.0-20190927141846-d69c422ba035
+	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/dustin/go-jsonpointer v0.0.0-20160814072949-ba0abeacc3dc // indirect
 	github.com/dustin/gojson v0.0.0-20160307161227-2e71ec9dd5ad // indirect
-	github.com/fogleman/gg v1.2.1-0.20190220221249-0403632d5b90
 	github.com/garyburd/go-oauth v0.0.0-20180319155456-bca2e7f09a17 // indirect
-	github.com/go-chi/chi/v5 v5.0.3
 	github.com/go-stack/stack v1.8.0 // indirect
 	github.com/gobwas/glob v0.2.3 // indirect
-	github.com/gocolly/colly v1.2.0
+	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0 // indirect
 	github.com/golang/groupcache v0.0.0-20191027212112-611e8accdfc9 // indirect
+	github.com/golang/protobuf v1.3.1 // indirect
 	github.com/gonum/floats v0.0.0-20181209220543-c233463c7e82 // indirect
 	github.com/gonum/internal v0.0.0-20181124074243-f884aa714029 // indirect
-	github.com/google/uuid v1.1.1
+	github.com/gorilla/websocket v1.4.2 // indirect
 	github.com/inconshreveable/log15 v0.0.0-20200109203555-b30bc20e4fd1 // indirect
-	github.com/itchyny/gojq v0.12.3
-	github.com/james-bowman/nlp v0.0.0-20191016091239-d9dbfaff30c6
+	github.com/itchyny/timefmt-go v0.1.2 // indirect
 	github.com/james-bowman/sparse v0.0.0-20190423065201-80c6877364c7 // indirect
-	github.com/jmoiron/sqlx v1.2.0
 	github.com/kennygrant/sanitize v1.2.4 // indirect
 	github.com/kevinburke/go-types v0.0.0-20200309064045-f2d4aea18a7a // indirect
 	github.com/kevinburke/go.uuid v1.2.0 // indirect
 	github.com/kevinburke/rest v0.0.0-20200429221318-0d2892b400f8 // indirect
-	github.com/kevinburke/twilio-go v0.0.0-20200424172635-4f0b2357b852
 	github.com/mattn/go-colorable v0.1.6 // indirect
-	github.com/mattn/go-sqlite3 v1.14.8
-	github.com/mmcdole/gofeed v1.0.0-beta2
+	github.com/mattn/go-isatty v0.0.12 // indirect
 	github.com/mmcdole/goxpp v0.0.0-20181012175147-0068e33feabf // indirect
-	github.com/nfnt/resize v0.0.0-20180221191011-83c6a9932646
-	github.com/olebedev/when v0.0.0-20190311101825-c3b538a97254
 	github.com/pkg/errors v0.9.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/robertkrimen/otto v0.0.0-20180617131154-15f95af6e78d // indirect
-	github.com/rs/zerolog v1.15.0
 	github.com/saintfish/chardet v0.0.0-20120816061221-3af4cd4741ca // indirect
-	github.com/slack-go/slack v0.9.4
 	github.com/spaolacci/murmur3 v1.1.0 // indirect
 	github.com/stretchr/objx v0.2.0 // indirect
-	github.com/stretchr/testify v1.4.0
 	github.com/temoto/robotstxt v1.1.1 // indirect
-	github.com/trubitsyn/go-zero-width v1.0.1
 	github.com/ttacon/builder v0.0.0-20170518171403-c099f663e1c2 // indirect
 	github.com/ttacon/libphonenumber v1.1.0 // indirect
-	github.com/velour/velour v0.0.0-20160303155839-8e090e68d158
-	golang.org/x/crypto v0.0.0-20211117183948-ae814b36b871
 	golang.org/x/exp v0.0.0-20191014171548-69215a2ee97e // indirect
+	golang.org/x/image v0.0.0-20190802002840-cff245a6509b // indirect
 	golang.org/x/net v0.0.0-20211118161319-6a13c67c3ce4 // indirect
 	golang.org/x/sync v0.0.0-20201020160332-67f06af15bc9 // indirect
 	golang.org/x/sys v0.0.0-20211117180635-dee7805ff2e1 // indirect
 	golang.org/x/text v0.3.7 // indirect
 	gonum.org/v1/gonum v0.6.0 // indirect
 	google.golang.org/appengine v1.6.5 // indirect
-	gopkg.in/go-playground/webhooks.v5 v5.13.0
 	gopkg.in/sourcemap.v1 v1.0.5 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
 )
