@@ -204,56 +204,56 @@ type fileResp struct {
 }
 
 type slackFile struct {
-	ID                 string        `json:"id"`
-	Created            int           `json:"created"`
-	Timestamp          int           `json:"timestamp"`
-	Name               string        `json:"name"`
-	Title              string        `json:"title"`
-	Mimetype           string        `json:"mimetype"`
-	Filetype           string        `json:"filetype"`
-	PrettyType         string        `json:"pretty_type"`
-	User               string        `json:"user"`
-	Editable           bool          `json:"editable"`
-	Size               int           `json:"size"`
-	Mode               string        `json:"mode"`
-	IsExternal         bool          `json:"is_external"`
-	ExternalType       string        `json:"external_type"`
-	IsPublic           bool          `json:"is_public"`
-	PublicURLShared    bool          `json:"public_url_shared"`
-	DisplayAsBot       bool          `json:"display_as_bot"`
-	Username           string        `json:"username"`
-	URLPrivate         string        `json:"url_private"`
-	URLPrivateDownload string        `json:"url_private_download"`
-	Thumb64            string        `json:"thumb_64"`
-	Thumb80            string        `json:"thumb_80"`
-	Thumb360           string        `json:"thumb_360"`
-	Thumb360W          int           `json:"thumb_360_w"`
-	Thumb360H          int           `json:"thumb_360_h"`
-	Thumb480           string        `json:"thumb_480,omitempty"`
-	Thumb480W          int           `json:"thumb_480_w,omitempty"`
-	Thumb480H          int           `json:"thumb_480_h,omitempty"`
-	Thumb160           string        `json:"thumb_160"`
-	Thumb720           string        `json:"thumb_720,omitempty"`
-	Thumb720W          int           `json:"thumb_720_w,omitempty"`
-	Thumb720H          int           `json:"thumb_720_h,omitempty"`
-	Thumb800           string        `json:"thumb_800,omitempty"`
-	Thumb800W          int           `json:"thumb_800_w,omitempty"`
-	Thumb800H          int           `json:"thumb_800_h,omitempty"`
-	Thumb960           string        `json:"thumb_960,omitempty"`
-	Thumb960W          int           `json:"thumb_960_w,omitempty"`
-	Thumb960H          int           `json:"thumb_960_h,omitempty"`
-	Thumb1024          string        `json:"thumb_1024,omitempty"`
-	Thumb1024W         int           `json:"thumb_1024_w,omitempty"`
-	Thumb1024H         int           `json:"thumb_1024_h,omitempty"`
-	ImageExifRotation  int           `json:"image_exif_rotation"`
-	OriginalW          int           `json:"original_w"`
-	OriginalH          int           `json:"original_h"`
-	Permalink          string        `json:"permalink"`
-	PermalinkPublic    string        `json:"permalink_public"`
-	Channels           []string      `json:"channels"`
-	Groups             []interface{} `json:"groups"`
-	Ims                []interface{} `json:"ims"`
-	CommentsCount      int           `json:"comments_count"`
+	ID                 string   `json:"id"`
+	Created            int      `json:"created"`
+	Timestamp          int      `json:"timestamp"`
+	Name               string   `json:"name"`
+	Title              string   `json:"title"`
+	Mimetype           string   `json:"mimetype"`
+	Filetype           string   `json:"filetype"`
+	PrettyType         string   `json:"pretty_type"`
+	User               string   `json:"user"`
+	Editable           bool     `json:"editable"`
+	Size               int      `json:"size"`
+	Mode               string   `json:"mode"`
+	IsExternal         bool     `json:"is_external"`
+	ExternalType       string   `json:"external_type"`
+	IsPublic           bool     `json:"is_public"`
+	PublicURLShared    bool     `json:"public_url_shared"`
+	DisplayAsBot       bool     `json:"display_as_bot"`
+	Username           string   `json:"username"`
+	URLPrivate         string   `json:"url_private"`
+	URLPrivateDownload string   `json:"url_private_download"`
+	Thumb64            string   `json:"thumb_64"`
+	Thumb80            string   `json:"thumb_80"`
+	Thumb360           string   `json:"thumb_360"`
+	Thumb360W          int      `json:"thumb_360_w"`
+	Thumb360H          int      `json:"thumb_360_h"`
+	Thumb480           string   `json:"thumb_480,omitempty"`
+	Thumb480W          int      `json:"thumb_480_w,omitempty"`
+	Thumb480H          int      `json:"thumb_480_h,omitempty"`
+	Thumb160           string   `json:"thumb_160"`
+	Thumb720           string   `json:"thumb_720,omitempty"`
+	Thumb720W          int      `json:"thumb_720_w,omitempty"`
+	Thumb720H          int      `json:"thumb_720_h,omitempty"`
+	Thumb800           string   `json:"thumb_800,omitempty"`
+	Thumb800W          int      `json:"thumb_800_w,omitempty"`
+	Thumb800H          int      `json:"thumb_800_h,omitempty"`
+	Thumb960           string   `json:"thumb_960,omitempty"`
+	Thumb960W          int      `json:"thumb_960_w,omitempty"`
+	Thumb960H          int      `json:"thumb_960_h,omitempty"`
+	Thumb1024          string   `json:"thumb_1024,omitempty"`
+	Thumb1024W         int      `json:"thumb_1024_w,omitempty"`
+	Thumb1024H         int      `json:"thumb_1024_h,omitempty"`
+	ImageExifRotation  int      `json:"image_exif_rotation"`
+	OriginalW          int      `json:"original_w"`
+	OriginalH          int      `json:"original_h"`
+	Permalink          string   `json:"permalink"`
+	PermalinkPublic    string   `json:"permalink_public"`
+	Channels           []string `json:"channels"`
+	Groups             []any    `json:"groups"`
+	Ims                []any    `json:"ims"`
+	CommentsCount      int      `json:"comments_count"`
 	InitialComment     struct {
 		ID        string `json:"id"`
 		Created   int    `json:"created"`
